+++ src/main/java/at/hannibal2/skyhanni/features/dungeon/CroesusUnopenedChestTracker.kt
@@ -1,0 +1,40 @@
+package at.hannibal2.skyhanni.features.dungeon
+
+import at.hannibal2.skyhanni.SkyHanniMod
+import at.hannibal2.skyhanni.events.GuiContainerEvent
+import at.hannibal2.skyhanni.utils.InventoryUtils
+import at.hannibal2.skyhanni.utils.ItemUtils.getLore
+import at.hannibal2.skyhanni.utils.LorenzColor
+import at.hannibal2.skyhanni.utils.LorenzUtils
+import at.hannibal2.skyhanni.utils.RenderUtils.highlight
+import net.minecraft.client.gui.inventory.GuiChest
+import net.minecraft.inventory.ContainerChest
+import net.minecraftforge.fml.common.eventhandler.EventPriority
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
+
+class CroesusUnopenedChestTracker {
+
+    @SubscribeEvent(priority = EventPriority.LOW)
+    fun onBackgroundDrawn(event: GuiContainerEvent.BackgroundDrawnEvent) {
+        if (!LorenzUtils.inSkyblock) return
+        if (!SkyHanniMod.feature.dungeon.croesusUnopenedChestTracker) return
+
+        if (event.gui !is GuiChest) return
+        val guiChest = event.gui
+        val chest = guiChest.inventorySlots as ContainerChest
+        val chestName = chest.lowerChestInventory.displayName.unformattedText.trim()
+
+        if (chestName == "Croesus") {
+            for (slot in InventoryUtils.getItemsInOpenChest()) {
+                val stack = slot.stack
+                val lore = stack.getLore()
+                if (lore.any { it.contains("Click to view") }) {
+                    if (!lore.any { it.contains("Chests have been opened!") }) {
+                        slot highlight LorenzColor.GREEN
+                    }
+                }
+            }
+        }
+
+    }
+}